+++ include/rainsprite_state.h
@@ -8,8 +8,8 @@ typedef struct {
     unsigned short sprite_num;
     unsigned char is_exploding;
     unsigned char has_landed;
-    unsigned int explosion_countdown;
+    unsigned short explosion_countdown;
 } RainspriteState;
 
-void rainsprite_state_init(RainspriteState * state, unsigned short sprite_num, unsigned int starting_x, unsigned int starting_y);
+void rainsprite_state_init(RainspriteState * state, unsigned short sprite_num, unsigned short starting_x, unsigned short starting_y);
 #endif
